DYNAP.ENH: Create BaseClass for dynap classes.
